+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ### next
+  * tpm2_createak: remove -k persistant option. Use tpm2_evictcontrol.
   * tpm2_nvincrement: -L and -F pcr policy options go away, replaced with pcr password minilanguage.
   * tpm2_nvwrite: -L and -F pcr policy options go away, replaced with pcr password minilanguage.
   * tpm2_nvread: -L and -F pcr policy options go away, replaced with pcr password minilanguage.